+++ HISTORY.md
@@ -7,6 +7,8 @@ This file will contain current and previous changes to Crackle. Make sure to upd
 
 ## 2025-11-29
  * docs: add hooking tip to API.md
+ * feat: hooking for menus, see API.md
+ * docs: added docs for menu hooking
 
 ## 2025-11-25
  * combine all files into one index.js file